Moved ElementSidebar into Sidebar dir
