+++ client/components/editor/Sidebar/ElementSidebar/index.vue
