+++ client/components/editor/Sidebar/Navigation.vue
