+++ client/components/editor/Sidebar/index.vue
